Move string tests to separate file
